+++ pkg/kubernetes/portforward.go
@@ -1,4 +1,4 @@
-package main
+package kubernetes
 
 import (
 	"context"
